+++ pytorch_lightning/trainer/trainer.py
@@ -14,6 +14,7 @@ from pytorch_lightning.callbacks import ModelCheckpoint, EarlyStopping, Callback
 from pytorch_lightning.core.lightning import LightningModule
 from pytorch_lightning.loggers import LightningLoggerBase
 from pytorch_lightning.profiler import SimpleProfiler, PassThroughProfiler, BaseProfiler
+from pytorch_lightning.trainer.seed import seed_everything
 from pytorch_lightning.trainer.auto_mix_precision import TrainerAMPMixin
 from pytorch_lightning.trainer.callback_config import TrainerCallbackConfigMixin
 from pytorch_lightning.trainer.callback_hook import TrainerCallbackHookMixin
@@ -32,8 +33,7 @@ from pytorch_lightning.trainer.training_loop import TrainerTrainLoopMixin
 from pytorch_lightning.trainer.training_tricks import TrainerTrainingTricksMixin
 from pytorch_lightning.trainer.lr_finder import TrainerLRFinderMixin
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities import rank_zero_warn
-from pytorch_lightning.utilities import parsing
+from pytorch_lightning.utilities import rank_zero_warn, parsing
 
 
 try:
@@ -124,12 +124,14 @@ class Trainer(
             num_sanity_val_steps: int = 5,
             truncated_bptt_steps: Optional[int] = None,
             resume_from_checkpoint: Optional[str] = None,
-            profiler: Optional[BaseProfiler] = None,
+            profiler: Optional[Union[BaseProfiler, bool]] = None,
             benchmark: bool = False,
+            deterministic: bool = False,
             reload_dataloaders_every_epoch: bool = False,
             auto_lr_find: Union[bool, str] = False,
             replace_sampler_ddp: bool = True,
             progress_bar_callback: Optional[Union[ProgressBarBase, bool]] = True,
+            terminate_on_nan: bool = False,
             auto_scale_batch_size: Optional[str] = None,
             amp_level: str = 'O1',  # backward compatible, todo: remove in v0.8.0
             default_save_path=None,  # backward compatible, todo: remove in v0.8.0
@@ -140,7 +142,6 @@ class Trainer(
             use_amp=None,  # backward compatible, todo: remove in v0.9.0
             show_progress_bar=None,  # backward compatible, todo: remove in v0.9.0
             nb_sanity_val_steps=None,  # backward compatible, todo: remove in v0.8.0
-            terminate_on_nan: bool = False,
             **kwargs
     ):
         r"""
@@ -293,6 +294,8 @@ class Trainer(
 
             benchmark: If true enables cudnn.benchmark.
 
+            deterministic: If true enables cudnn.deterministic
+
             terminate_on_nan: If set to True, will terminate training (by raising a `ValueError`) at the
                 end of each training batch, if any of the parameters or the loss are NaN or +/-inf.
 
@@ -302,6 +305,13 @@ class Trainer(
                 Additionally, can be set to either `power` that estimates the batch size through
                 a power search or `binsearch` that estimates the batch size through a binary search.
         """
+
+        self.deterministic = deterministic
+        torch.backends.cudnn.deterministic = self.deterministic
+        if self.deterministic:
+            # fixing non-deterministic part of horovod
+            # https://github.com/PyTorchLightning/pytorch-lightning/pull/1572/files#r420279383
+            os.environ["HOROVOD_FUSION_THRESHOLD"] = str(0)
 
         # Init callbacks
         self.callbacks = callbacks or []
@@ -574,6 +584,7 @@ class Trainer(
              ('process_position', (<class 'int'>,), 0),
              ('profiler',
               (<class 'pytorch_lightning.profiler.profilers.BaseProfiler'>,
+               <class 'bool'>,
                <class 'NoneType'>),
               None),
              ...
@@ -612,6 +623,33 @@ class Trainer(
 
         Only arguments of the allowed types (str, float, int, bool) will
         extend the `parent_parser`.
+
+        Examples:
+            >>> import argparse
+            >>> import pprint
+            >>> parser = argparse.ArgumentParser()
+            >>> parser = Trainer.add_argparse_args(parser)
+            >>> args = parser.parse_args([])
+            >>> pprint.pprint(vars(args))  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
+            {...
+             'check_val_every_n_epoch': 1,
+             'checkpoint_callback': True,
+             'default_root_dir': None,
+             'deterministic': False,
+             'distributed_backend': None,
+             'early_stop_callback': False,
+             ...
+             'logger': True,
+             'max_epochs': 1000,
+             'max_steps': None,
+             'min_epochs': 1,
+             'min_steps': None,
+             ...
+             'profiler': None,
+             'progress_bar_callback': True,
+             'progress_bar_refresh_rate': 1,
+             ...}
+
         """
         parser = ArgumentParser(parents=[parent_parser], add_help=False, )